1) add subject hashes as field in courses and make two separate collections (courses and subjects)
2) change professor and subject reviews

+++ entity/subject.go
@@ -8,15 +8,15 @@ import (
 
 // Subject describes a subject (example: SMA0356 - Cálculo IV)
 type Subject struct {
-	Code          string             `firestore:"code,omitempty"`
-	Name          string             `firestore:"name,omitempty"`
-	Description   string             `firestore:"desc,omitempty"`
-	ClassCredits  int                `firestore:"class,omitempty"`
-	AssignCredits int                `firestore:"assign,omitempty"`
-	TotalHours    string             `firestore:"hours,omitempty"`
-	Requirements  []string           `firestore:"requirements,omitempty"`
-	Optional      bool               `firestore:"optional,omitempty"`
-	Stats         map[string]float64 `firestore:"stats,omitempty"`
+	Code          string         `firestore:"code,omitempty"`
+	Name          string         `firestore:"name,omitempty"`
+	Description   string         `firestore:"desc,omitempty"`
+	ClassCredits  int            `firestore:"class,omitempty"`
+	AssignCredits int            `firestore:"assign,omitempty"`
+	TotalHours    string         `firestore:"hours,omitempty"`
+	Requirements  []string       `firestore:"requirements,omitempty"`
+	Optional      bool           `firestore:"optional,omitempty"`
+	Stats         map[string]int `firestore:"stats,omitempty"`
 }
 
 func (s Subject) Hash() string {
